Data JPA Rest endpoint demonstration
This commit adds a simple JPA entity and a repository using an embedded
in-memory database. Spring Boot provides "starters" for the most common
use cases: for instance, adding `spring-boot-starter-data-jpa` will get
you running with Spring Data JPA and Hibernate: Spring Boot will scan
entities from the package where the `@SpringBootApplication` is defined
and create any Spring Data repository it finds the same way.

Adding `spring-boot-starter-data-rest` will expose any of your
repositories as REST endpoints. If you try
`http://localhost:8080/speakers` this app, you'll see some entries and
others endpoints you can browse to.

There are some decisions Spring Boot can't take for the users. There
isn't a sensible default database so one must be added. Adding H2 is
enough for Spring Boot to automatically create an in-memory database.

Finally, this commit uses an Hibernate feature that's looking for a file
called `import.sql` at the root of the classpath. If such file is
present it is processed automatically. This allows us to create a set of
speakers automatically on startup.