Spring Security demonstration
If we add the `spring-boot-starter-security` starter, our app is secured
using basic auth and a single user named `user` with an auto-generated
password that is printed on startup. That's rather annoying and we'd
like to use a defined set of users.

One way to do that is to define our own `UserDetailsManager` bean. This
commit adds a simple in-memory one with two users: user has the `USER`
role and hero has the `USER` and `HERO` roles.