Devtools demonstration
Spring Boot provides development tools that are particularly interesting
when working in your IDE. If you add `spring-boot-devtools`, the
application is automatically tuned for local development:

* All template engine caches will be disabled.
* If H2 is present the h2 console will be exposed at `/h2-console`.
* A LiveReload server will be started as part of the app, allowing you
to connect to your app using the free LiveReload browser plugin.

More importantly, Spring Boot watches your files and every time
the classpath changes it either refreshes the browser (via LiveReload)
or restart the application and refresh the browser when it's done.

Concretely, if you are working on a template engine or a static resource
and that one changes, the browser will refresh automatically. IF you
change anything else (code, configuration, etc), the application will
restart to apply those changes.

If you try to create an error page in `static/error/404.html`, you'll
notice any unmapped URI will redirect to that page. We could also add
a custom health indicator and show that `/health` shows it once we've
build the project.